requirements.txt
docetl>=0.2.6
python-dotenv>=1.0.0
pyyaml>=6.0
beautifulsoup4>=4.12
lxml>=4.9
pymupdf>=1.23
pdfplumber>=0.10
pandas>=2.0
requests>=2.31
pytest>=7.4
rapidfuzz>=3.0
tqdm>=4.66
tiktoken>=0.5
